api/requirements.txt
Django==3.2.6
djangorestframework
django-cors-headers
psycopg2-binary
redis
huey
xlwt
redis-server
pandas
JayDeBeApi
openpyxl
xlrd